requirements.txt
Flask==2.2.5
gspread==5.8.0
google-auth
python-dotenv
gunicorn

